SystemBuilder: better handling of system creation with non-default kind while deferred
kind_id() checks for existing phases and removes them, but the addition of the default
OnUpdate phase has not been processed yet by that point (if deferred).

Special case for existing entities being passed in, because they effectively have the same
issue; the system builder cannot reliably know whether a kind is going to be set on the
existing entity in the next defer flush.